+++ app/includes/swap.tpl
@@ -1,10 +1,7 @@
 <!-- Swap Page -->
-<main class="tab-pane swap-tab active" ng-if="globalService.currentTab ===globalService.tabs.swap.id"
+<main class="tab-pane swap-tab active" ng-if="globalService.currentTab === globalService.tabs.swap.id"
       ng-controller='swapCtrl' ng-cloak>
-
     <div ng-switch="stage">
-
-
         <div ng-switch-default>
 
             @@if (site === 'cew' ) { @@include( '../includes/swap-stage-1.tpl', { "site": "cew" } ) }
@@ -23,6 +20,25 @@
     </div>
 
     <section class="changenow-contact text-center">
+        <div ng-show="orderResult && orderResult.id" ng-init="showDetails = false;" class="marg-v-md">
+
+            <div class="d-flex align-items-center justify-content-center">
+                <button class="btn" ng-click="showDetails = !showDetails;">
+                    Swap Details
+                </button>
+            </div>
+            <div ng-show="showDetails">
+                <label for="orderResult">
+                    Order
+                    <a href="{{'https://changenow.io/exchange/txs/' + orderResult.id}}" target="_blank" rel="noopener">
+                        {{orderResult.id}}
+                    </a>
+                </label>
+                <textarea readonly class="form-control" name="orderResult" id="orderResult" rows="8">
+                {{orderResult | json}}
+            </textarea>
+            </div>
+        </div>
         <p><a class="btn-warning btn-sm"
               href={{mailHref()}}
               target="_blank" rel="noopener"> Issue with your Swap? Contact support</a></p>
