+++ app/scripts/ethFuncs.js
@@ -1,5 +1,4 @@
 "use strict";
-const { Contract } = require("./contract");
 
 var ethFuncs = function() {};
 
@@ -19,7 +18,7 @@ ethFuncs.isChecksumAddress = function(address) {
     return address === ethUtil.toChecksumAddress(address);
 };
 ethFuncs.validateHexString = function(str) {
-    if (typeof str === "undefined") return false;
+    if (str == null) return false;
     else if (str === "") return true;
     str =
         str.substring(0, 2) === "0x"
@@ -91,149 +90,6 @@ ethFuncs.getFunctionSignature = function(name) {
         .slice(0, 8);
 };
 
-/*
-    returns <Promise> {data, msg, error: false}
- */
-
-function mapToGasEst(tx) {
-    return Object.assign({}, tx, {
-        value: new BigNumber(tx.value).toNumber()
-    });
-}
-
-ethFuncs.estimateGas = function(dataObj, notifyError = true) {
-    return new Promise((resolve, reject) => {
-        const tx = mapToGasEst(dataObj);
-
-        ajaxReq.getEstimatedGas(tx, function(data) {
-            if (data.error || parseInt(data.data) === -1) {
-                notifyError && uiFuncs.notifier.danger(data);
-
-                reject(data);
-            } else {
-                resolve(new BigNumber(data.data).toNumber());
-            }
-        });
-    });
-};
-
-/*
-
-    Given function, contract, and tx data, generates data and sends call, returns decoded outputs
-    @param string | contract.abi[n] _func
-    @param Contract contract
-    @param Transaction tx
-    @returns Promise<{error: bool, data: []any}>
- */
-
-ethFuncs.call = function(
-    _func,
-    contract,
-    {
-        network = ajaxReq.type,
-        inputs = null,
-        from = null,
-        value = 0,
-        unit = "ether"
-    } = {}
-) {
-    return new Promise((resolve, reject) => {
-        const { node } = contract;
-
-        const { tx: transObj, _function, error } = ethFuncs.prepContractData(
-            _func,
-            contract,
-            {
-                inputs,
-                from,
-                value,
-                unit
-            }
-        );
-
-        if (error) {
-            reject({ error: transObj, data: null });
-        } else {
-            // if reading from contract, send call
-
-            node.lib.getEthCall(
-                { to: transObj.to, data: transObj.data },
-                function(data) {
-                    if (data.error) {
-                        reject(data);
-                    } else
-                        resolve(
-                            Object.assign({}, data, {
-                                data: ethFuncs.decodeOutputs(_function, data)
-                            })
-                        );
-                }
-            );
-        }
-    });
-};
-
-/*
-
-    Estimate gasPrice of tx to contract
-
-    sent over contract's set network
-
-    @param string | contract.abi[n[ _func
-    @param Contract contract
-    @param Tx transaction
-
-    @returns tx: Tx {gasLimit: }
-
- */
-
-ethFuncs.estGasContract = function(
-    _func,
-    contract,
-    {
-        network = ajaxReq.type,
-        inputs = null,
-        from = null,
-        value = 0,
-        unit = "ether"
-    } = {}
-) {
-    return new Promise((resolve, reject) => {
-        const tx = { network, inputs, from, value, unit };
-
-        const { error, tx: _tx } = ethFuncs.prepContractData(
-            _func,
-            contract,
-            tx
-        );
-
-        if (error) {
-            reject(error);
-        } else {
-            Object.assign(tx, _tx);
-
-            const estObj = mapToGasEst({
-                from: tx.from,
-                data: tx.data,
-                to: contract.address,
-                value: etherUnits.toWei(tx.value, tx.unit)
-            });
-
-            contract.node.lib.getEstimatedGas(estObj, function(data) {
-                if (data.error || parseInt(data.data) === -1) {
-                    reject(Object.assign({}, data, { error: true }));
-                } else {
-                    resolve(
-                        Object.assign({}, tx, {
-                            gasLimit: adjustGas(data.data)
-                        })
-                    );
-                }
-            });
-        }
-    });
-};
-
 // utils
 
 ethFuncs.encodeInputs = function encodeInputs({ inputs }) {
@@ -268,63 +124,6 @@ ethFuncs.decodeOutputs = function decodeOutputs(contractFunction, data) {
 
         return i;
     });
-};
-
-/*
-
-
-    Generates tx data from contract function
-
-    @param string  | contract.abi[n] _FUNCTION
-    @param Contract contract
-    @param Tx {}
-    @returns {error: bool | error, {tx: Tx, _function: contract.abi.function} } if cannot estimate gas
-
- */
-
-ethFuncs.prepContractData = function(
-    _FUNCTION,
-    contract,
-    { inputs = [], from, value = 0, unit = "ether" } = {}
-) {
-    const ERROR = { error: true, tx: null, _function: null };
-
-    if (!(contract instanceof Contract)) {
-        return ERROR;
-    }
-
-    let _function = null;
-
-    if (typeof _FUNCTION === "string") {
-        _function = contract.abi.find(itm => itm.name === _FUNCTION);
-    } else {
-        _function = _FUNCTION;
-    }
-
-    if (!contract.validFunction(_function)) return ERROR;
-
-    _function.inputs.forEach((item, i) => (item.value = inputs[i] || ""));
-
-    let data = ethFuncs.getFunctionSignature(
-        ethUtil.solidityUtils.transformToFullName(_function)
-    );
-
-    if (!data) {
-        return ERROR;
-    }
-
-    const inputs__ = ethFuncs.encodeInputs(_function);
-
-    return {
-        tx: {
-            to: contract.address,
-            data: ethFuncs.sanitizeHex(data + inputs__),
-            value,
-            unit
-        },
-        _function,
-        error: null
-    };
 };
 
 module.exports = ethFuncs;
